made my e2e testing into unit tests (waiting for other features to be done)
